[type: refactor] renamed logging plugin  to loggingConsole plugin
