+++ pom.xml
@@ -123,6 +123,8 @@
         <nacos-client.version>2.0.0</nacos-client.version>
         <spring-security.version>5.3.10.RELEASE</spring-security.version>
         <grpc.version>1.33.1</grpc.version>
+        <rocketmq-client.version>4.9.2</rocketmq-client.version>
+        <lz4-java.version>1.8.0</lz4-java.version>
         <!--maven plugin version-->
         <exec-maven-plugin.version>1.6.0</exec-maven-plugin.version>
         <jacoco-maven-plugin.version>0.8.7</jacoco-maven-plugin.version>
