--- shenyu-plugin/shenyu-plugin-logging/src/test/java/org/apache/shenyu/plugin/logging/LoggingPluginTest.java
@@ -1,102 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.shenyu.plugin.logging;
-
-import org.apache.shenyu.common.dto.RuleData;
-import org.apache.shenyu.common.dto.SelectorData;
-import org.apache.shenyu.common.enums.PluginEnum;
-import org.apache.shenyu.plugin.api.ShenyuPluginChain;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.mock.http.server.reactive.MockServerHttpRequest;
-import org.springframework.mock.web.server.MockServerWebExchange;
-import org.springframework.web.server.ServerWebExchange;
-import reactor.core.publisher.Mono;
-import reactor.test.StepVerifier;
-
-import java.net.InetSocketAddress;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.spy;
-import static org.mockito.Mockito.when;
-
-/**
- * The Test Case For DebugPlugin.
- */
-@ExtendWith(MockitoExtension.class)
-public final class LoggingPluginTest {
-
-    private LoggingPlugin loggingPlugin;
-
-    private ServerWebExchange exchange;
-
-    private RuleData ruleData;
-
-    private ShenyuPluginChain chain;
-
-    private SelectorData selectorData;
-
-    @BeforeEach
-    public void setUp() {
-        this.loggingPlugin = new LoggingPlugin();
-        this.ruleData = mock(RuleData.class);
-        this.chain = mock(ShenyuPluginChain.class);
-        this.selectorData = mock(SelectorData.class);
-        MockServerHttpRequest request = MockServerHttpRequest
-                .get("localhost")
-                .remoteAddress(new InetSocketAddress(8090))
-                .header("X-source", "mock test")
-                .queryParam("queryParam", "Hello,World")
-                .build();
-        this.exchange = spy(MockServerWebExchange.from(request));
-    }
-
-    @Test
-    public void testDoExecute() {
-        ServerWebExchange.Builder builder = mock(ServerWebExchange.Builder.class);
-        when(exchange.mutate()).thenReturn(builder);
-        when(builder.request(any(LoggingPlugin.LoggingServerHttpRequest.class))).thenReturn(builder);
-        when(builder.response(any(LoggingPlugin.LoggingServerHttpResponse.class))).thenReturn(builder);
-        when(builder.build()).thenReturn(exchange);
-        when(chain.execute(any())).thenReturn(Mono.empty());
-        Mono<Void> result = loggingPlugin.doExecute(exchange, chain, selectorData, ruleData);
-        // Sorry, I do not how to mock this case by an simply way, so I give up.
-
-        StepVerifier.create(result).expectSubscription().verifyComplete();
-    }
-
-    @Test
-    public void testGetOrder() {
-        assertEquals(loggingPlugin.getOrder(), PluginEnum.LOGGING.getCode());
-    }
-
-    @Test
-    public void testNamed() {
-        assertEquals(loggingPlugin.named(), PluginEnum.LOGGING.getName());
-    }
-
-    @Test
-    public void testSkip() {
-        assertFalse(loggingPlugin.skip(exchange));
-    }
-}
