+++ shenyu-spring-boot-starter/shenyu-spring-boot-starter-plugin/pom.xml
@@ -44,7 +44,7 @@
         <module>shenyu-spring-boot-starter-plugin-context-path</module>
         <module>shenyu-spring-boot-starter-plugin-grpc</module>
         <module>shenyu-spring-boot-starter-plugin-redirect</module>
-        <module>shenyu-spring-boot-starter-plugin-logging</module>
+        <module>shenyu-spring-boot-starter-plugin-logging-console</module>
         <module>shenyu-spring-boot-starter-plugin-modify-response</module>
         <module>shenyu-spring-boot-starter-plugin-request</module>
         <module>shenyu-spring-boot-starter-plugin-oauth2</module>
