+++ shenyu-spring-boot-starter/shenyu-spring-boot-starter-plugin/shenyu-spring-boot-starter-plugin-logging-console/src/main/resources/META-INF/spring.factories
@@ -16,4 +16,4 @@
 #
 
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-org.apache.shenyu.springboot.starter.plugin.logging.console.rocketmq.LoggingRocketMQPluginConfiguration
+org.apache.shenyu.springboot.starter.plugin.logging.console.LoggingConsolePluginConfiguration
