+++ shenyu-spring-boot-starter/shenyu-spring-boot-starter-plugin/shenyu-spring-boot-starter-plugin-logging-console/src/main/resources/META-INF/spring.provides
@@ -15,5 +15,4 @@
 # limitations under the License.
 #
 
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-org.apache.shenyu.springboot.starter.plugin.logging.console.rocketmq.LoggingRocketMQPluginConfiguration
+provides: shenyu-spring-boot-starter-plugin-logging-console
